+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## v18.0.0-SNAPSHOT (in development / unreleased)
+## v19.0.0-SNAPSHOT (in development / unreleased)
+
+## v18.0.0-rcx
 
 ### 🎁 New Features
 
@@ -14,11 +16,13 @@
   settings from its containing `Form`.
   * `HoistInput` has been moved into the `input` package to clarify that these are lower level controls and independent
   of the Forms package.
+  * `RestGrid` now supports a `mask` prop.  RestGrid loading is now masked by default.   
+  * `Chart` component now supports a built-in zoomout gesture:  Drag right-to-left on charts with x-axis zooming.
 
 ### 💥 Breaking Changes
 * Major changes to Form (see above).  `HoistInput` imports will also need to be adjusted to move from `form` to `input`.
 
-## v17.0.0 - 2018-12-21
+## v17.0.0 
 
 ### 💥 Breaking Changes